+++ NEWS.md
@@ -89,5 +89,5 @@
 * Now the functions "iterate_more" and "alt_gsmar" also warn about near-unit-roots and return results from all the estimation rounds.
 * Summary and print methods now display the number of parameters and observations.
 * Updated some of the documentation
-* Added argument "close_connections" to functions fitGSMAR. Set to FALSE when calling the functions from Rmarkdown to prevent an error. 
+* fitGSMAR does not call closeAllConnections() on exit anymore; instead, it only closes the connections it opened.
 * Fixed the argument "precission" in profile_logliks to "precision".